refactor

+++ Frelance.Application/Mediatr/Commands/Projects/UpdateProjectCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Frelance.Application.Mediatr.Commands.Projects;
 
-public record UpdateProjectCommand(int Id, string Title, string Description, DateTime Deadline, List<string> Technologies, float Budget) : IRequest<Unit>;
+public record UpdateProjectCommand(int Id, string Title, string Description, DateTime Deadline, List<string> Technologies, decimal Budget) : IRequest<Unit>;
